Bug fixed empirical sleep time and added array per channel
Corrections were made and removal of empirical times, responsible for freezing the IOC.
When using the internal buffer mode, a single Array was used for all channels,
It is necessary to process the array later, now the channels are saved in separate Arrays and timestamps.
The epics-in-docker submodule was added for IOC compilation via docker.